Move avatar helper functions into PrimerHelpers
